backend/requirements.txt
# LLM Interview Simulator - Python Dependencies (AWS Edition)
# ==========================================================
# This file lists all the Python packages needed to run our FastAPI backend.
# Install these with: pip install -r requirements.txt
#
# Why each package is needed:
# - fastapi: The main web framework for building our API
# - uvicorn: ASGI server to run our FastAPI application
# - python-dotenv: Load environment variables from .env files
# - pydantic: Data validation and serialization for our API models
# - python-multipart: Handle file uploads and form data
# - gunicorn: Production WSGI server (alternative to uvicorn)
# - boto3: AWS SDK for Python (Bedrock, Polly, Transcribe, S3, RDS)
# - sqlalchemy: Database ORM for PostgreSQL
# - psycopg2: PostgreSQL adapter for Python
# - cryptography: Security library (required for SQLAlchemy with PostgreSQL)

# Core FastAPI framework and server
fastapi==0.104.1              # Modern, fast web framework for building APIs
uvicorn[standard]==0.24.0      # ASGI server with additional features (websockets, etc.)

# Environment and configuration
python-dotenv==1.0.0          # Load environment variables from .env files

# Data validation and serialization
pydantic==2.5.0               # Data validation using Python type annotations
pydantic-settings==2.1.0      # Settings management for Pydantic

# File handling and forms
python-multipart==0.0.6       # Handle multipart/form-data (file uploads, forms)

# Production server
gunicorn==21.2.0              # WSGI HTTP Server for production deployment

# AWS SDK and services
boto3==1.29.7                 # AWS SDK for Python (Bedrock, Polly, Transcribe, S3, etc.)

# Database (PostgreSQL)
sqlalchemy==2.0.23            # SQL Toolkit and ORM
psycopg2-binary==2.9.9       # PostgreSQL adapter
alembic==1.12.1               # Database migration tool

# Security and encryption
cryptography==41.0.7         # Cryptographic library for database encryption

# HTTP client
httpx==0.25.2                # Async HTTP client for API calls

# CORS and middleware
#python-cors==1.0.0            # CORS middleware (if needed beyond FastAPI)
requests==2.32.3